refactor theme structure
