+++ lib/src/controls/title_bar.dart
@@ -4,8 +4,9 @@ import 'dart:ui';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart' hide BoxDecoration, BoxShadow;
 import 'package:flutter_inset_shadow/flutter_inset_shadow.dart';
-import 'package:window_manager/window_manager.dart' hide TitleBarStyle;
+import 'package:window_manager/window_manager.dart';
 import 'package:xp_ui/src/styles/colors.dart';
+import 'package:xp_ui/src/styles/theme/titlebar_theme.dart';
 import 'package:xp_ui/xp_ui.dart';
 
 class TitleBar extends StatefulWidget implements PreferredSizeWidget {
@@ -40,7 +41,10 @@ class TitleBar extends StatefulWidget implements PreferredSizeWidget {
         canDragWindow = true;
 
   const TitleBar.dialog(this.title,
-      {super.key, this.showCloseButton = true, this.showHelpButton = false, this.onHelpButtonPressed})
+      {super.key,
+      this.showCloseButton = true,
+      this.showHelpButton = false,
+      this.onHelpButtonPressed})
       : canCloseWindow = false,
         canDragWindow = false,
         showMaximizeButton = false,
@@ -54,7 +58,8 @@ class TitleBar extends StatefulWidget implements PreferredSizeWidget {
 }
 
 class _TitleBarState extends State<TitleBar> with WindowListener {
-  bool get _isDesktop => Platform.isLinux || Platform.isMacOS || Platform.isWindows;
+  bool get _isDesktop =>
+      Platform.isLinux || Platform.isMacOS || Platform.isWindows;
   bool isMaximized = false;
 
   @override
@@ -83,7 +88,7 @@ class _TitleBarState extends State<TitleBar> with WindowListener {
 
   @override
   Widget build(BuildContext context) {
-    final style = XpTheme.of(context).titleBarStyle;
+    final style = XpTheme.of(context).titleBarTheme;
     final HSLColor color = HSLColor.fromColor(style.backgroundColor);
     final Widget titleWidget = Row(
       mainAxisSize: MainAxisSize.min,
@@ -93,7 +98,10 @@ class _TitleBarState extends State<TitleBar> with WindowListener {
         ),
         Text(
           widget.title,
-          style: TextStyle(color: style.foregroundColor, fontSize: 13, fontFamily: 'Trebuchet'),
+          style: TextStyle(
+              color: style.foregroundColor,
+              fontSize: 13,
+              fontFamily: 'Trebuchet'),
         ),
       ],
     );
@@ -149,7 +157,9 @@ class _TitleBarState extends State<TitleBar> with WindowListener {
                   ),
                 if (widget.showMaximizeButton)
                   TitleBarActionButton(
-                    icon: !isMaximized ? ActionButtonIcon.maximize : ActionButtonIcon.maximized,
+                    icon: !isMaximized
+                        ? ActionButtonIcon.maximize
+                        : ActionButtonIcon.maximized,
                     onPressed: () async {
                       if (!_isDesktop) return;
                       if (!isMaximized) {
@@ -166,7 +176,8 @@ class _TitleBarState extends State<TitleBar> with WindowListener {
                         : () {
                             if (!_isDesktop) return;
                             // SystemNavigator.pop();
-                            SystemChannels.platform.invokeMethod('System.exitApplication', {
+                            SystemChannels.platform
+                                .invokeMethod('System.exitApplication', {
                               'type': AppExitType.required.name,
                               'exitCode': exitCode,
                             });
@@ -202,7 +213,10 @@ class _TitleBarButtonHoverState extends State<_TitleBarButtonHover> {
         },
         child: DecoratedBox(
           decoration: BoxDecoration(
-              gradient: _hover ? const RadialGradient(colors: [Color(0x6AFFFFFF), XpColors.transparent]) : null),
+              gradient: _hover
+                  ? const RadialGradient(
+                      colors: [Color(0x6AFFFFFF), XpColors.transparent])
+                  : null),
           child: widget.child,
         ));
   }
@@ -220,7 +234,7 @@ class _XpCloseButtonState extends State<XpCloseButton> {
   bool _tap = false;
   @override
   Widget build(BuildContext context) {
-    final TitleBarStyle titleBarStyle = XpTheme.of(context).titleBarStyle;
+    final TitleBarThemeData titleBarStyle = XpTheme.of(context).titleBarTheme;
     return Padding(
       padding: const EdgeInsets.only(left: 2),
       child: GestureDetector(
@@ -255,12 +269,23 @@ class _XpCloseButtonState extends State<XpCloseButton> {
                     offset: const Offset(1, 2),
                     color: _tap ? XpColors.darkRed : XpColors.lightRed,
                     blurRadius: 1),
-                const BoxShadow(inset: true, offset: Offset(-2, 2), color: XpColors.red, blurRadius: 1),
-                const BoxShadow(inset: true, offset: Offset(2, -2), color: XpColors.red, blurRadius: 1)
+                const BoxShadow(
+                    inset: true,
+                    offset: Offset(-2, 2),
+                    color: XpColors.red,
+                    blurRadius: 1),
+                const BoxShadow(
+                    inset: true,
+                    offset: Offset(2, -2),
+                    color: XpColors.red,
+                    blurRadius: 1)
               ],
               gradient: _tap
-                  ? const LinearGradient(colors: [XpColors.red, XpColors.darkRed])
-                  : const LinearGradient(colors: [XpColors.lightRed, XpColors.red], begin: Alignment.topLeft)),
+                  ? const LinearGradient(
+                      colors: [XpColors.red, XpColors.darkRed])
+                  : const LinearGradient(
+                      colors: [XpColors.lightRed, XpColors.red],
+                      begin: Alignment.topLeft)),
           child: _TitleBarButtonHover(
             child: Image.asset(
               ActionButtonIcon.close.assetPath,
@@ -287,7 +312,7 @@ class _TitleBarActionButtonState extends State<TitleBarActionButton> {
 
   @override
   Widget build(BuildContext context) {
-    final TitleBarStyle titleBarStyle = XpTheme.of(context).titleBarStyle;
+    final TitleBarThemeData titleBarStyle = XpTheme.of(context).titleBarTheme;
     return Padding(
       padding: const EdgeInsets.only(left: 2),
       child: GestureDetector(
@@ -312,25 +337,39 @@ class _TitleBarActionButtonState extends State<TitleBarActionButton> {
                 BoxShadow(
                     inset: true,
                     offset: const Offset(-1, 1),
-                    color: _tap ? titleBarStyle.backgroundShade5 : titleBarStyle.backgroundShade2,
+                    color: _tap
+                        ? titleBarStyle.backgroundShade5
+                        : titleBarStyle.backgroundShade2,
                     blurRadius: 1),
                 BoxShadow(
                     inset: true,
                     offset: const Offset(1, 2),
-                    color: _tap ? titleBarStyle.backgroundShade5 : titleBarStyle.backgroundShade2,
+                    color: _tap
+                        ? titleBarStyle.backgroundShade5
+                        : titleBarStyle.backgroundShade2,
                     blurRadius: 1),
                 BoxShadow(
                     inset: true,
                     offset: const Offset(-2, 2),
-                    color: _tap ? titleBarStyle.backgroundShade3 : titleBarStyle.backgroundColor,
+                    color: _tap
+                        ? titleBarStyle.backgroundShade3
+                        : titleBarStyle.backgroundColor,
                     blurRadius: 1),
-                BoxShadow(inset: true, offset: const Offset(2, -2), color: titleBarStyle.backgroundColor, blurRadius: 1)
+                BoxShadow(
+                    inset: true,
+                    offset: const Offset(2, -2),
+                    color: titleBarStyle.backgroundColor,
+                    blurRadius: 1)
               ],
               gradient: _tap
-                  ? LinearGradient(colors: [titleBarStyle.backgroundShade3, titleBarStyle.backgroundShade5])
-                  : LinearGradient(
-                      colors: [titleBarStyle.backgroundShade1, titleBarStyle.backgroundShade3],
-                      begin: Alignment.topLeft)),
+                  ? LinearGradient(colors: [
+                      titleBarStyle.backgroundShade3,
+                      titleBarStyle.backgroundShade5
+                    ])
+                  : LinearGradient(colors: [
+                      titleBarStyle.backgroundShade1,
+                      titleBarStyle.backgroundShade3
+                    ], begin: Alignment.topLeft)),
           child: _TitleBarButtonHover(
             child: Image.asset(
               widget.icon.assetPath,
