+++ lib/src/styles/theme/progressbar_theme.dart
@@ -1,0 +1,14 @@
+import 'dart:ui';
+
+import 'package:xp_ui/src/styles/colors.dart';
+
+class ProgressBarThemeData {
+  final Color backgroundColor;
+  final Color trackColor;
+  final Color borderColor;
+
+  const ProgressBarThemeData(
+      {this.backgroundColor = XpColors.white,
+      this.trackColor = XpColors.green,
+      this.borderColor = XpDefaultThemeColors.outLineColor});
+}
