+++ lib/xp_ui.dart
@@ -3,7 +3,7 @@ library xp_ui;
 export './xp_ui.dart';
 export './src/controls/xp_app.dart';
 export './src/controls/button.dart';
-export './src/styles/theme.dart';
+export 'src/styles/theme/theme.dart';
 export './src/indicators/progress.dart';
 export './src/controls/title_bar.dart';
 export './src/controls/textbox.dart';
